backend/requirements.txt
fastapi
uvicorn[standard]
passlib
bcrypt==3.2.0

pydantic[email]           
motor
python-dotenv
axios
sqlalchemy
python-multipart
python-jose
PyJWT[crypto]
pandas
scipy
joblib
scikit-learn
pymongo 
axios
certifi

